refactor: features allowing compiling sans rocksdb
This adds features allowing us to compile interdependent common modules
sans rocksdb to make their builds quicker

There should be no functionality changes for this refactor
