+++ common/fs/Cargo.toml
@@ -11,7 +11,7 @@ integration_tests = []
 #local
 http = { package = "mz-http", path = "../http" }
 metrics = { package = "metrics", path = "../metrics" }
-state = { package = "state", path = "../state" }
+state = { package = "state", path = "../state", default-features=false }
 notify_stream = { package = "notify_stream", path = "../notify_stream" }
 
 #error
