+++ common/state/Cargo.toml
@@ -7,12 +7,14 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [features]
+default = ["state"]
 integration_tests = []
+state = ["rocksdb"]
 
 [dependencies]
 derivative = "2.2"
 futures = "0.3"
-rocksdb = "0.19"
+rocksdb = { version = "0.19", optional = true }
 async-channel = "1.6"
 thiserror = "1.0"
 tracing = "0.1"
